Datatypes

Description

Module Datatypes bevat de nieuwe versie van alle data om mee te testen en debuggen. Deze waardes vervangen in het verloop van het project de data in DatatypesOld. Op dit moment werkt deze data helaas nog niet met de functies uit Lib.

Documentation

newtype Product

Constructors

Product String

newtype Transactie

Constructors

Transactie (Set Product)

newtype Transacties

Constructors

Transacties [Transactie]

data KeuzeOutput

Constructors

Waarde Double

WaardeProduct (Double, Product)

Foutief String

prodA :: Product

Product om mee te debuggen

prodB :: Product

Product om mee te debuggen

prodC :: Product

Product om mee te debuggen

prodD :: Product

Product om mee te debuggen

prodE :: Product

Product om mee te debuggen

prodF :: Product

Product dat niet voorkomt in de transacties om mee te debuggen

transAB :: Transactie

Subset van een transactie om mee te debuggen

transDB :: Transactie

Subset van een transactie om mee te debuggen

transT1 :: Transactie

Transactie om mee te debuggen

transT2 :: Transactie

Transactie om mee te debuggen

transT3 :: Transactie

Transactie om mee te debuggen

alleTrans :: Transacties

Lijst van Transactie om mee te debuggen

Deze variabele moet een dataset simuleren.