DatatypesOld

Description

Module DatatypesOld bevat de oude versie van alle data om mee te testen en debuggen. Deze waardes worden in het verloop van het project vervangen voor de data in Datatypes.

Documentation

setA :: Set [Char]

Product om mee te debuggen

setB :: Set [Char]

Product om mee te debuggen

setC :: Set [Char]

Product om mee te debuggen

setD :: Set [Char]

Product om mee te debuggen

setE :: Set [Char]

Product om mee te debuggen

setF :: Set [Char]

Product dat niet voorkomt in de transacties om mee te debuggen

setG :: Set a

Geen product om mee te debuggen

setAB :: Set [Char]

Subset van een transactie om mee te debuggen

setDB :: Set [Char]

Subset van een transactie om mee te debuggen

setT1 :: Set [Char]

Transactie om mee te debuggen

setT2 :: Set [Char]

Transactie om mee te debuggen

setT3 :: Set [Char]

Transactie om mee te debuggen

listTrans :: [Set [Char]]

Transacties om mee te debuggen

Deze variabele moet een dataset simuleren.