Lib

Description

Module Lib bevat de belangrijkste functies van aHaskellPriori.

Documentation

bestandLezen :: FilePath -> IO ()

Leest een bestand in en print iedere regel

Credits: https://stackoverflow.com/questions/24234517/understanding-withfile-with-example

support :: Transactie -> Transacties -> Double

Berekend de support van één Transactie: Support(X)

De Transactie X moet één of meerdere Product bevatten. Mocht Transactie X leeg zijn, dan is de support gelijk aan 0.0

confidence :: Product -> Product -> Transacties -> Double

Berekend de confidence van een Product ten opzichte van een ander Product: Confidence(X -> Y)

Mocht een van de Producten leeg zijn, dan is de confidence 0.0. Mocht de support van Product X 0.0 zijn, dan is de confidence tevens 0.0. Op deze manier worden zero division errors en NaN-waardes volkomen.

lift :: Product -> Product -> Transacties -> Double

Berekend de lift van een Product ten opzichte van een ander Product: Lift(X -> Y)

Mocht de support van Product Y 0.0 zijn, dan is de confidence tevens 0.0. Op deze manier worden zero division errors en NaN-waardes volkomen.

bestLift :: Product -> Transacties -> (Double, Product)

Berekend het beste Product Y voor de formule: Lift(X -> Y)

Ondanks dat meerdere producten een even hoge lift kunnen hebben met setX zal alleen de laatste (alfabetisch gesoorteerde) waarde worden weergegeven.

bestSupport :: Transacties -> (Double, Product)

Berekend het hoogste Product X voor de formule: Support(X)

Ondanks dat meerdere producten een even hoge support kunnen hebben zal alleen de laatste (alfabetisch gesoorteerde) waarde worden weergegeven.