Matrix

Description

Module Matrix bevat alle code van de oude opzet voor het Apriori algoritme. Na verloop van tijd bleek het per keer berekenen van de support, confidence en lift snel genoeg. Het van te voren berekenen van de waardes in de verschillende matrixen blijkt dus niet nodig.

Documentation

setLen :: Set a -> Int

Returned de lengte van een set

emptySupportMatrix :: Num b => Set a -> Matrix b

Genereerd een leeg matrix met de lengtes van het aantal unieke producten in een dataset