Tools

Description

Module Tools bevat alle hulpmiddelen om de hoofdfuncties in Lib te laten werken.

Documentation

recursiveUnion :: Ord a => [Set a] -> Set a

Fold een lijst met transacties zodat alle unieke producten overblijven

Fold was eerst zelf uitgeschreven: recursiveUnion (x:xs) = x `union` recursiveUnion xs

setToListOfSets :: Ord a => Set a -> [Set a]

Veranderd een set van elementen naar een gesoorteerde lijst van (Set) singletons

floatDiv :: Int -> Int -> Double

Deelt twee integers en geeft een float terug

Credits: stackoverflow.com/questions/3275193/whats-the-right-way-to-divide-two-int-values-to-obtain-a-float

listOfTuplesMax :: Ord a => [(a, b)] -> (a, b)

Returned een tuple met de maximale first uit een lijst met tuples

Credits: stackoverflow.com/questions/52910840/haskell-finding-maximum-value-in-a-list-of-tuples