Tools

Description

Module Tools bevat alle hulpmiddelen om de hoofdfuncties in Lib te laten werken.

Documentation

recursiveUnion :: Transacties -> Transactie

Fold Transacties naar Transactie zodat alle unieke producten overblijven

Fold was eerst zelf uitgeschreven: recursiveUnion (x:xs) = x `union` recursiveUnion xs

setToListOfSets :: Transactie -> [Product]

Veranderd Transactie naar een gesoorteerde lijst van Producten

floatDiv :: Int -> Int -> Double

Deelt twee integers en geeft een float terug

Credits: stackoverflow.com/questions/3275193/whats-the-right-way-to-divide-two-int-values-to-obtain-a-float

listOfTuplesMax :: Ord a => [(a, b)] -> (a, b)

Returned een tuple met de maximale first uit een lijst met tuples

Credits: stackoverflow.com/questions/52910840/haskell-finding-maximum-value-in-a-list-of-tuples

vereniging :: Transactie -> Transactie -> Transactie

Returned de vereniging van twee Transactie

intersectie :: Transactie -> Transactie -> Transactie

Returned de intersectie van twee Transactie